Add "managed contact lists" option; fixed big bug
I added an option to have PMP-CC add/remove members from only the lists
that have been assigned to membership levels while leaving other lists
alone. This allows the site to stay in sync with CC while keeping a
standard CC workflow for other lists.
Additionally there was a GIGANTIC bug where any modifications to the
profile page for a member would cause the member to be added to EVERY CC
mailing list. That's no longer possible with this version.